make PrivilegedVmap module-private
