+++ litebox_platform_lvbs/src/lib.rs
@@ -453,61 +453,10 @@ type UserMutPtr<T> =
 /// Type-level marker for the VTL0 physical-pointer provider.
 pub enum Vmap {}
 
-/// Type-level marker for MSHV operations authorized to modify protected VTL0 frames.
-pub(crate) struct PrivilegedVmap;
-
 impl<const ALIGN: usize> GlobalVmapManager<ALIGN> for Vmap {
     type Manager = crate::host::LvbsLinuxKernel;
     fn manager() -> &'static Self::Manager {
         crate::platform_low()
-    }
-}
-
-impl<const ALIGN: usize> GlobalVmapManager<ALIGN> for PrivilegedVmap {
-    type Manager = PrivilegedVmap;
-    fn manager() -> &'static Self::Manager {
-        &PrivilegedVmap
-    }
-}
-
-unsafe impl<const ALIGN: usize> VmapManager<ALIGN> for PrivilegedVmap {
-    type MapInfo = LvbsPhysPageMapInfo;
-
-    unsafe fn vmap(
-        &self,
-        pages: &PhysPageAddrArray<ALIGN>,
-        perms: PhysPageMapPermissions,
-    ) -> Result<Self::MapInfo, PhysPointerError> {
-        // SAFETY: callers uphold the raw mapping contract. This provider is used only
-        // for independently authorized HEKI patch and ring-buffer writes.
-        unsafe { crate::platform_low().vmap_privileged(pages, perms) }
-    }
-
-    unsafe fn vunmap(
-        &self,
-        map_info: Self::MapInfo,
-    ) -> Result<(), (PhysPointerError, Self::MapInfo)> {
-        // SAFETY: `map_info` came from the same LVBS platform mapper and has no outstanding uses
-        // beyond the physical-pointer guard that is dropping it.
-        unsafe {
-            <crate::host::LvbsLinuxKernel as VmapManager<ALIGN>>::vunmap(
-                crate::platform_low(),
-                map_info,
-            )
-        }
-    }
-
-    fn validate_unowned(&self, pages: &PhysPageAddrArray<ALIGN>) -> Result<(), PhysPointerError> {
-        crate::platform_low().validate_unowned(pages)
-    }
-
-    unsafe fn protect(
-        &self,
-        pages: &PhysPageAddrArray<ALIGN>,
-        perms: PhysPageMapPermissions,
-    ) -> Result<(), PhysPointerError> {
-        // SAFETY: callers uphold `VmapManager::protect`; this forwards unchanged to LVBS.
-        unsafe { crate::platform_low().protect(pages, perms) }
     }
 }
